Atomic Type Casting Without Stripping Attributes

Atomic type casting in R is generally performed using the functions as.logical, as.integer, as.double, as.character, as.complex, and as.raw.

Converting an object between atomic types using these functions strips the object of its attributes, including attributes such as names and dimensions.

The functions provided here by the 'tinycodet' package preserve all attributes - except the "class" attribute.

The functions are as follows:

as_bool(): converts object to atomic type logical (TRUE, FALSE, NA).

as_int(): converts object to atomic type integer.

as_dbl(): converts object to atomic type double (AKA decimal numbers).

as_chr(): converts object to atomic type character.

as_cplx(): converts object to atomic type complex.

as_raw():converts object to atomic type raw.

Usage

as_bool(x, ...)

as_int(x, ...)

as_dbl(x, ...)

as_chr(x, ...)

as_cplx(x, ...)

as_raw(x, ...)

Arguments

x

vector, matrix, array (or a similar object where all elements share the same type).

...

further arguments passed to or from other methods.

Value

The converted object.

See also

Examples

x <- c(rep(0, 2), seq(0, 2.5, by=0.5)) |> matrix(ncol=2)
colnames(x) <- c("one", "two")
attr(x, "test") <- "test"
print(x)
#>      one two
#> [1,] 0.0 1.0
#> [2,] 0.0 1.5
#> [3,] 0.0 2.0
#> [4,] 0.5 2.5
#> attr(,"test")
#> [1] "test"

# notice that in all following, attributes (except class) are conserved:
as_bool(x)
#>        one  two
#> [1,] FALSE TRUE
#> [2,] FALSE TRUE
#> [3,] FALSE TRUE
#> [4,]  TRUE TRUE
#> attr(,"test")
#> [1] "test"
as_int(x)
#>      one two
#> [1,]   0   1
#> [2,]   0   1
#> [3,]   0   2
#> [4,]   0   2
#> attr(,"test")
#> [1] "test"
as_dbl(x)
#>      one two
#> [1,] 0.0 1.0
#> [2,] 0.0 1.5
#> [3,] 0.0 2.0
#> [4,] 0.5 2.5
#> attr(,"test")
#> [1] "test"
as_chr(x)
#>      one   two  
#> [1,] "0"   "1"  
#> [2,] "0"   "1.5"
#> [3,] "0"   "2"  
#> [4,] "0.5" "2.5"
#> attr(,"test")
#> [1] "test"