Construct Formula With Safer Environment Specification

When creating a formula with the tilde ( ~ ) operator, and storing a formula in a variable to be used later, the environment is captured by the formula.
Therefore, any object in the captured environment might not be freed from the memory, potentially creating some memory leak (see also the Examples section below).

The form() function is a convenience function, to create and return/store a formula more safely, by having no default environment, and allowing the user to specify the environment explicitly.
It can also convert a single string to a formula, again allowing for explicit environment specification, and having no environment by default.

Usage

form(f, env = NULL)

Arguments

f

either one of the following:

a literal single string, or a variable containing a single string.
If a string is given that does not contain a tilde ( ~ ), a tilde is prepended to the string before converting it to a formula.
Do not forget to surrounded special names with back-ticks.

a literal formula.
Note that if a literal formula is given, it must actually be a literal formula, not a variable that contains a formula, as that would defeat the whole point of the form() function.

env

environment of the formula.
Defaults to NULL.

Value

A formula with no environment, or a formula with the specified environment.

See also

Examples

# basic examples ====

(myform <- form(a ~ b))
#> a ~ b
#> NULL
(myform2 <- form("a ~ b"))
#> a ~ b
#> NULL
mystring <- "a ~ b"
(myform3 <- form(mystring))
#> a ~ b
#> NULL
form("a")
#> ~a
#> NULL

myform <- form(x ~ y)
environment(myform) # NULL
#> NULL
mydata <- data.frame(x = rnorm(1e5), y = rnorm(1e5))
lm(myform, data = mydata)
#> 
#> Call:
#> lm(formula = myform, data = mydata)
#> 
#> Coefficients:
#> (Intercept)            y  
#>   -0.003002    -0.002401  
#> 

#############################################################################

# showcasing environment capture ====
# see also http://adv-r.had.co.nz/memory.html

f1 <- function() {
  foo <- c(letters, LETTERS)
  return(10)
}
x1 <- f1()
environment(x1) |> as.list() # empty, since no formula is used inside f1(), so safe
#> list()

f2 <- function() {
  foo <- c(letters, LETTERS)
  out <- a ~ b
  return(out)
}
x2 <- f2()
environment(x2) |> as.list() # NOT safe: contains all objects from f2()
#> $out
#> a ~ b
#> <environment: 0x000001fa9484b580>
#> 
#> $foo
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z" "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L"
#> [39] "M" "N" "O" "P" "Q" "R" "S" "T" "U" "V" "W" "X" "Y" "Z"
#> 
exists("foo", envir = environment(x2)) # = TRUE: "foo" still exists!
#> [1] TRUE
environment(x2)$foo # can still access it; probably won't be removed by gc()
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z" "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L"
#> [39] "M" "N" "O" "P" "Q" "R" "S" "T" "U" "V" "W" "X" "Y" "Z"

f3 <- function() {
  foo <- c(letters, LETTERS)
  out <- form(a ~ b)
  return(out)
}
x3 <- f3()
environment(x3) |> as.list() # empty, since form() is used, so safe
#> list()

f4 <- function() {
  foo <- c(letters, LETTERS)
  out <- form("a ~ b")
  return(out)
}
x4 <- f4()
environment(x4) |> as.list() # empty, since form() is used, so safe
#> list()