Report infix operators

The report_inops() function returns a data.frame listing the infix operators defined in the current environment (like the global environment, or the environment within a function), or a user specified environment. It also reports from which packages the infix operators came from.

Usage

report_inops(env)

Arguments

env

an optional environment to give, where the function should look for infix operators.
When not specified, the current environment (like the global environment, or the environment within a function) is used.

Value

A data.frame.

See also

Examples


report_inops()
#> NULL

import_inops("stringi")
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done

report_inops()
#>    infix_operator package
#> 1           %s!=% stringi
#> 2          %s!==% stringi
#> 3            %s$% stringi
#> 4            %s*% stringi
#> 5            %s+% stringi
#> 6            %s<% stringi
#> 7           %s<=% stringi
#> 8           %s==% stringi
#> 9          %s===% stringi
#> 10           %s>% stringi
#> 11          %s>=% stringi
#> 12       %stri!=% stringi
#> 13      %stri!==% stringi
#> 14        %stri$% stringi
#> 15        %stri*% stringi
#> 16        %stri+% stringi
#> 17        %stri<% stringi
#> 18       %stri<=% stringi
#> 19       %stri==% stringi
#> 20      %stri===% stringi
#> 21        %stri>% stringi
#> 22       %stri>=% stringi