Load R-package and its Re-exports and/or its (Reverse) Dependencies Under a Single Alias

The import_as() function imports the namespace of an R-package, and optionally also its re-exports, dependencies, and extensions, all under the same alias. The specified alias will be placed in the current environment (like the global environment, or the environment within a function).

Usage

import_as(
  alias,
  main_package,
  re_exports = TRUE,
  dependencies = NULL,
  extensions = NULL,
  lib.loc = .libPaths(),
  loadorder = c("dependencies", "main_package", "extensions")
)

Arguments

alias

a syntactically valid non-hidden name giving the alias object where the package(s) are to be loaded into.
This name can be given either as a single string (i.e. "alias."), or as a one-sided formula with a single term (i.e. ~ alias.).

main_package

a single string, giving the name of the main package to load under the given alias.

re_exports

TRUE or FALSE.

If re_exports = TRUE the re-exports from the main_package are added to the alias together with the main package. This is the default, as it is analogous to the behaviour of base R's :: operator.

If re_exports = FALSE, these re-exports are not added together with the main package. The user can still load the packages from which the re-exported functions came from, by specifying them in the dependencies argument.

dependencies

an optional character vector, giving the names of the dependencies of the main_package to be loaded also under the alias.
Defaults to NULL, which means no dependencies are loaded.
See pkg_get_deps to quickly get dependencies from a package.

extensions

an optional character vector, giving the names of the extensions of the main_package to be loaded also under the alias.
Defaults to NULL, which means no extensions are loaded.

lib.loc

character vector specifying library search path (the location of R library trees to search through).
The lib.loc argument would usually be .libPaths().
See also loadNamespace.

loadorder

the character vector
c("dependencies", "main_package", "extensions"),
or some re-ordering of this character vector, giving the relative load order of the groups of packages.
See Details section for more information.

Value

A locked environment object, similar to the output of loadNamespace, with the name as specified in the alias argument, will be created.

This object, referred to as the "(package) alias object", will contain the exported functions from the specified package(s).

The alias object will be placed in the current environment (like the global environment, or the environment within a function).

To use, for example, function "some_function()" from alias "alias.", use:

alias.$some_function()

To see the special attributes of this alias object, use attr.import.

To "unload" the package alias object, simply remove it (i.e. rm(list="alias.")).

Details

Expanded Definitions of Some Arguments

"Re-exports" are functions that are defined in the dependencies of the main_package, but are re-exported in the namespace of the main_package.

"Dependencies" are here defined as any R-package appearing in the "Depends", "Imports", or "LinkingTo" fields of the Description file of the main_package. So no recursive dependencies.

"Extensions" are here defined as direct reverse-depends or direct reverse-imports. It does not matter if these are CRAN or non-CRAN packages. However, the intended meaning of an extension is not merely a reverse dependency, but a package that actually extends the functionality of the main_package.

Why Aliasing Multiple Packages is Useful
To use an R-package with its extension packages or dependencies, whilst avoiding the disadvantages of attaching a package (see tinycodet_import), one would traditionally use the :: operator like so:

main_package::some_function1()
extension1::some_function2()
extension2::some_function3()

This becomes cumbersome as more packages are needed and/or as the package name(s) become longer.
The import_as() function avoids this issue by allowing multiple related packages to be loaded under a single alias, allowing one to code like this:

import_as(
   ~ alias., "main_package",
   extensions = c("extension1", "extension2"),
   lib.loc = .libPaths()
)
alias.$some_function1()
alias.$some_function2()
alias.$some_function3()

Thus loading a package, or multiple directly related packages, under a single alias, which import_as() provides, avoids the above issues. Loading (a) package(s) under an alias is known as "aliasing" (a) package(s).

Notice that the import_as() function has the lib.loc argument, allowing to specify the library path, which the :: operator does not directly provide.

Alias Naming Recommendation
To keep package alias object names easily distinguishable from other objects that can also be subset with the $ operator, I recommend ending (not starting!) all alias names with a dot (.) or underscore (_).

Regarding the Load Order
The order of the character vector given in the dependencies and extensions arguments matters. If multiple packages share objects with the same name, the objects of the package named last will overwrite those of the earlier named packages.

The loadorder argument defaults to the character vector
c("dependencies", "main_package", "extensions"),
which is the recommended setting.
This setting results in the following load order:

The dependencies, in the order specified by the depenencies argument.

The main_package (see argument main_package), including re-exports (if re_exports = TRUE).

The extensions, in the order specified by the extensions argument.

Other Details
The import_as() function does not support loading base/core R under an alias.

See also

Examples


import_as( # this creates the 'tdt.' object
  "tdt.", "data.table", extensions = "tidytable"
)
#> Importing packages...
#> Done
#> You can now access the functions using `tdt.$`
#> Methods will work like normally 
#> For conflicts report, packages order, and other attributes, run `attr.import(tdt.)` 
# same as:
import_as(
  ~ tdt., "data.table", extensions = "tidytable"
)
#> Importing packages...
#> Done
#> You can now access the functions using `tdt.$`
#> Methods will work like normally 
#> For conflicts report, packages order, and other attributes, run `attr.import(tdt.)`