(Un)Expose Infix Operators From Package Namespace in the Current Environment

import_inops(expose=...) exposes infix operators specified in a package or an alias object to the current environment (like the global environment or the environment within a function).

import_inops(unexpose=...) "unexposes" (i.e. removes) the infix operators specified in a package or an alias object from the current environment (like the global environment or the environment within a function).
Note that in this case only infix operators exposed by the 'tinycodet' import system will be removed from the current environment; "regular" (i.e. user-specified) infix operators will not be touched.

Usage

import_inops(expose = NULL, unexpose = NULL, lib.loc = .libPaths(), ...)

Arguments

expose, unexpose

either one of the following:

an alias object as produced by the import_as function.

a string giving the package name.

lib.loc

character vector specifying library search path (the location of R library trees to search through).
Only used when supplying a string to expose / unexpose, and ignored when supplying an alias object to expose / unexpose (the library is path already stored inside the alias object).
The lib.loc argument would usually be .libPaths().
See also loadNamespace.

...

additional arguments, only relevant if the expose argument is used.
See import_inops.control.

Value

If using argument expose:

The infix operators specified in the given package or alias will be placed in the current environment (like the Global environment, or the environment within a function).

If using argument unexpose:

The infix operators specified in the given package or alias, exposed by import_inops(), will be removed from the current environment (like the Global environment, or the environment within a function).

If such infix operators could not be found, this function simply returns NULL.

Details

Why Exposing Infix Operators Is Useful
To use a function from an R-package, while avoiding the disadvantages of attaching a package (see tinycodet_import), one would traditionally use the :: operator like so:

packagename::function_name()

This is, however, cumbersome with infix operators, as it forces one to code like this:

packagename::`%op%`(x,y)

Exposing infix operators to the current environment, using the import_inops() function, allows one to use infix operators without using cumbersome code, and without having to attach the infix operators globally.

Other Details
The import_inops() function does not support overloading base/core R operators.

When using import_inops() to remove infix operators from the current environment, it will use the attributes of those operators to determine if the infix operator came from the 'tinycodet' import system or not. Only infix operators exposed by the 'tinycodet' import system will be removed.

See also

Examples


import_inops(expose = "stringi") # expose infix operators from package
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done
import_inops(unexpose = "stringi") # remove the exposed infix operators from environment
#> checking for infix operators exposed to the current environment by `import_inops()` ...
#> Removing the following infix operators:
#> %s!=%, %s!==%, %s$%, %s*%, %s+%, %s<%, %s<=%, %s==%, %s===%, %s>%, %s>=%, %stri!=%, %stri!==%, %stri$%, %stri*%, %stri+%, %stri<%, %stri<=%, %stri==%, %stri===%, %stri>%, %stri>=%
#> Done

import_as(~ stri., "stringi")
#> Importing packages and registering methods...
#> Done
#> You can now access the functions using `stri.$`
#> For conflicts report, packages order, and other attributes, run `attr.import(stri.)` 
import_inops(expose = stri.) # expose infix operators from alias
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done
import_inops(unexpose = stri.) # unexposed infix operators from current environment
#> unexposing infix operators...
#> Done


# additional arguments (only used when exposing, not unexposing):
import_inops(expose = "stringi", exclude = "%s==%")
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done
import_inops(unexpose = "stringi")
#> checking for infix operators exposed to the current environment by `import_inops()` ...
#> Removing the following infix operators:
#> %s!=%, %s!==%, %s$%, %s*%, %s+%, %s<%, %s<=%, %s===%, %s>%, %s>=%, %stri!=%, %stri!==%, %stri$%, %stri*%, %stri+%, %stri<%, %stri<=%, %stri==%, %stri===%, %stri>%, %stri>=%
#> Done
import_inops(expose = "stringi", overwrite = FALSE)
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done
import_inops(unexpose = "stringi")
#> checking for infix operators exposed to the current environment by `import_inops()` ...
#> Removing the following infix operators:
#> %s!=%, %s!==%, %s$%, %s*%, %s+%, %s<%, %s<=%, %s==%, %s===%, %s>%, %s>=%, %stri!=%, %stri!==%, %stri$%, %stri*%, %stri+%, %stri<%, %stri<=%, %stri==%, %stri===%, %stri>%, %stri>=%
#> Done

import_as(~ stri., "stringi")
#> Importing packages and registering methods...
#> Done
#> You can now access the functions using `stri.$`
#> For conflicts report, packages order, and other attributes, run `attr.import(stri.)` 
import_inops(expose = stri., include.only = "%s==%")
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done
import_inops(unexpose = stri.)
#> unexposing infix operators...
#> Done
import_inops(expose = stri., overwrite = FALSE)
#> Checking for conflicting infix operators in the current environment...
#> Placing infix operators in current environment...
#> Done
import_inops(unexpose = stri.)
#> unexposing infix operators...
#> Done